codeforces solution

+++ 1335A - Candies and Two Sisters.cpp
@@ -5,10 +5,9 @@ int main() {
     short t;
     cin >> t;
     while (t--) {
-        long int n;
+         int n;
         cin >> n;
         cout << (n < 3 ? 0 : (n / 2 - (n % 2 == 0))) << endl;
     }
     return 0;
 }
-
